Add Vitest and distance tests
