+++ src/components/Benefits.tsx
@@ -1,4 +1,5 @@
 import React, { useRef, useState, useEffect } from 'react';
+import { calculateDistance } from '../utils/calculateDistance';
 import { motion, useInView, AnimatePresence } from 'framer-motion';
 import { ShieldCheck, Clock, Zap, Droplet, Wrench, CreditCard, ArrowRight, X, ChevronRight, MapPin, Navigation, AlertTriangle } from 'lucide-react';
 
@@ -49,20 +50,6 @@ const Benefits: React.FC = () => {
       window.removeEventListener('resize', checkMobile);
     };
   }, []);
-
-  // Calcular distância entre dois pontos em km (fórmula de Haversine)
-  const calculateDistance = (lat1: number, lon1: number, lat2: number, lon2: number): number => {
-    const R = 6371; // Raio da Terra em km
-    const dLat = (lat2 - lat1) * Math.PI / 180;
-    const dLon = (lon2 - lon1) * Math.PI / 180;
-    const a = 
-      Math.sin(dLat/2) * Math.sin(dLat/2) +
-      Math.cos(lat1 * Math.PI / 180) * Math.cos(lat2 * Math.PI / 180) * 
-      Math.sin(dLon/2) * Math.sin(dLon/2); 
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a)); 
-    const distance = R * c; // Distância em km
-    return Math.round(distance * 10) / 10; // Arredondar para uma casa decimal
-  };
 
   // Obter localização do usuário
   const getUserLocation = () => {
